Move StreamLayerClientCacheTest into separate file
Remove unused #includes.

Relates-to: OLPEDGE-720
