+++ olp-cpp-sdk-dataservice-write/tests/StreamLayerClientTest.cpp
@@ -20,8 +20,6 @@
 #include <gmock/gmock.h>
 #include <matchers/NetworkUrlMatchers.h>
 #include <mocks/NetworkMock.h>
-#include <olp/authentication/TokenProvider.h>
-#include <olp/core/cache/DefaultCache.h>
 #include <olp/core/client/ApiError.h>
 #include <olp/core/client/HRN.h>
 #include <olp/core/client/HttpResponse.h>
@@ -34,7 +32,6 @@
 #include <boost/uuid/uuid.hpp>
 #include <boost/uuid/uuid_generators.hpp>
 #include <boost/uuid/uuid_io.hpp>
-#include <testutils/CustomParameters.hpp>
 #include "HttpResponses.h"
 
 namespace {
@@ -604,548 +601,6 @@ TEST_F(StreamLayerClientTest, SDIIConcurrentPublishSameIngestApi) {
   async3.get();
   async4.get();
   async5.get();
-}
-
-class StreamLayerClientCacheMockTest : public StreamLayerClientTest {
- protected:
-  virtual std::shared_ptr<StreamLayerClient> CreateStreamLayerClient()
-      override {
-    olp::client::OlpClientSettings client_settings;
-
-    disk_cache_ = std::make_shared<olp::cache::DefaultCache>();
-    EXPECT_EQ(disk_cache_->Open(),
-              olp::cache::DefaultCache::StorageOpenResult::Success);
-
-    network_ = std::make_shared<NetworkMock>();
-    client_settings.network_request_handler = network_;
-    SetUpCommonNetworkMockCalls(*network_);
-
-    return std::make_shared<StreamLayerClient>(
-        olp::client::HRN{GetTestCatalog()}, client_settings, disk_cache_,
-        flush_settings_);
-  }
-
-  virtual void TearDown() override {
-    if (disk_cache_) {
-      disk_cache_->Close();
-    }
-    StreamLayerClientTest::TearDown();
-  }
-
-  void FlushDataOnSettingSuccessAssertions(
-      const boost::optional<int>& max_events_per_flush = boost::none) {
-    for (int i = 0; i < 5; i++) {
-      data_->push_back(' ');
-      data_->push_back(i);
-      auto error = client_->Queue(
-          PublishDataRequest().WithData(data_).WithLayerId(GetTestLayer()));
-      EXPECT_FALSE(error) << error.get();
-    }
-    auto response = client_->Flush().GetFuture().get();
-    if (!max_events_per_flush || *max_events_per_flush > 5) {
-      EXPECT_EQ(5, response.size());
-    } else if (*max_events_per_flush <= 0) {
-      EXPECT_TRUE(response.empty());
-    } else {
-      EXPECT_EQ(*max_events_per_flush, response.size());
-    }
-
-    for (auto& single_response : response) {
-      EXPECT_NO_FATAL_FAILURE(PublishDataSuccessAssertions(single_response));
-    }
-  }
-  void MaximumRequestsSuccessAssertions(const int maximum_requests,
-                                        int num_requests = 0) {
-    std::string expected_error = "Maximum number of requests has reached";
-    if (num_requests) {
-      if (num_requests > maximum_requests) {
-        EXPECT_NO_FATAL_FAILURE(QueueMultipleEvents(maximum_requests));
-        while (num_requests > maximum_requests) {
-          auto error = client_->Queue(
-              PublishDataRequest().WithData(data_).WithLayerId(GetTestLayer()));
-          EXPECT_TRUE(error);
-          ASSERT_EQ(expected_error.compare(*error), 0);
-          num_requests--;
-        }
-      }
-    } else if (maximum_requests) {
-      EXPECT_NO_FATAL_FAILURE(QueueMultipleEvents(maximum_requests));
-      auto error = client_->Queue(
-          PublishDataRequest().WithData(data_).WithLayerId(GetTestLayer()));
-      EXPECT_TRUE(error);
-      ASSERT_EQ(expected_error.compare(*error), 0);
-    }
-  }
-
-  std::shared_ptr<olp::cache::DefaultCache> disk_cache_;
-  FlushSettings flush_settings_;
-};
-
-TEST_F(StreamLayerClientCacheMockTest, FlushDataSingle) {
-  {
-    testing::InSequence dummy;
-
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_LOOKUP_INGEST), _, _, _, _))
-        .Times(1);
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_LOOKUP_CONFIG), _, _, _, _))
-        .Times(1);
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_GET_CATALOG), _, _, _, _))
-        .Times(1);
-    EXPECT_CALL(*network_, Send(IsPostRequest(URL_INGEST_DATA), _, _, _, _))
-        .Times(1);
-  }
-
-  auto error = client_->Queue(
-      PublishDataRequest().WithData(data_).WithLayerId(GetTestLayer()));
-
-  ASSERT_FALSE(error) << error.get();
-
-  auto response = client_->Flush().GetFuture().get();
-
-  ASSERT_FALSE(response.empty());
-  ASSERT_NO_FATAL_FAILURE(PublishDataSuccessAssertions(response[0]));
-}
-
-TEST_F(StreamLayerClientCacheMockTest, FlushDataMultiple) {
-  {
-    testing::InSequence dummy;
-
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_LOOKUP_INGEST), _, _, _, _))
-        .Times(1);
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_LOOKUP_CONFIG), _, _, _, _))
-        .Times(1);
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_GET_CATALOG), _, _, _, _))
-        .Times(1);
-    EXPECT_CALL(*network_, Send(IsPostRequest(URL_INGEST_DATA), _, _, _, _))
-        .Times(5);
-  }
-
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(5));
-
-  auto response = client_->Flush().GetFuture().get();
-
-  ASSERT_EQ(5, response.size());
-  for (auto& single_response : response) {
-    ASSERT_NO_FATAL_FAILURE(PublishDataSuccessAssertions(single_response));
-  }
-}
-
-TEST_F(StreamLayerClientCacheMockTest, DISABLED_FlushDataCancel) {
-  auto wait_for_cancel = std::make_shared<std::promise<void>>();
-  auto pause_for_cancel = std::make_shared<std::promise<void>>();
-
-  olp::http::RequestId request_id;
-  NetworkCallback send_mock;
-  CancelCallback cancel_mock;
-
-  std::tie(request_id, send_mock, cancel_mock) = GenerateNetworkMockActions(
-      wait_for_cancel, pause_for_cancel, {200, HTTP_RESPONSE_LOOKUP_CONFIG});
-
-  {
-    testing::InSequence dummy;
-
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_LOOKUP_INGEST), _, _, _, _))
-        .Times(1);
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_LOOKUP_CONFIG), _, _, _, _))
-        .Times(1)
-        .WillOnce(testing::Invoke(std::move(send_mock)));
-    EXPECT_CALL(*network_, Cancel(request_id))
-        .WillOnce(testing::Invoke(std::move(cancel_mock)));
-  }
-
-  auto error = client_->Queue(
-      PublishDataRequest().WithData(data_).WithLayerId(GetTestLayer()));
-
-  ASSERT_FALSE(error) << error.get();
-
-  auto promise = client_->Flush();
-  wait_for_cancel->get_future().get();
-  promise.GetCancellationToken().cancel();
-  pause_for_cancel->set_value();
-
-  auto response = promise.GetFuture().get();
-
-  ASSERT_EQ(1, response.size());
-
-  ASSERT_NO_FATAL_FAILURE(PublishFailureAssertions(response[0]));
-}
-
-TEST_F(StreamLayerClientCacheMockTest, FlushListenerMetrics) {
-  disk_cache_->Close();
-  flush_settings_.auto_flush_num_events = 3;
-  client_ = CreateStreamLayerClient();
-  {
-    testing::InSequence dummy;
-
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_LOOKUP_INGEST), _, _, _, _))
-        .Times(1);
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_LOOKUP_CONFIG), _, _, _, _))
-        .Times(1);
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_GET_CATALOG), _, _, _, _))
-        .Times(1);
-    EXPECT_CALL(*network_, Send(IsPostRequest(URL_INGEST_DATA), _, _, _, _))
-        .Times(3);
-  }
-
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(3));
-
-  auto default_listener = StreamLayerClient::DefaultListener();
-
-  client_->Enable(default_listener);
-
-  for (int i = 0; default_listener->GetNumFlushEvents() < 1; i++) {
-    std::this_thread::sleep_for(std::chrono::milliseconds(50));
-    if (i > 200) {
-      FAIL() << "Timeout waiting for Flush Event Listener Results";
-    }
-  }
-
-  EXPECT_EQ(1, default_listener->GetNumFlushEvents());
-  EXPECT_EQ(1, default_listener->GetNumFlushEventsAttempted());
-  EXPECT_EQ(0, default_listener->GetNumFlushEventsFailed());
-  EXPECT_EQ(3, default_listener->GetNumFlushedRequests());
-  EXPECT_EQ(0, default_listener->GetNumFlushedRequestsFailed());
-}
-
-TEST_F(StreamLayerClientCacheMockTest,
-       FlushListenerMetricsSetListenerBeforeQueuing) {
-  disk_cache_->Close();
-  flush_settings_.auto_flush_num_events = 3;
-  client_ = CreateStreamLayerClient();
-
-  auto default_listener = StreamLayerClient::DefaultListener();
-  client_->Enable(default_listener);
-  {
-    testing::InSequence dummy;
-
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_LOOKUP_INGEST), _, _, _, _))
-        .Times(1);
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_LOOKUP_CONFIG), _, _, _, _))
-        .Times(1);
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_GET_CATALOG), _, _, _, _))
-        .Times(1);
-    EXPECT_CALL(*network_, Send(IsPostRequest(URL_INGEST_DATA), _, _, _, _))
-        .Times(3);
-  }
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(3));
-
-  for (int i = 0; default_listener->GetNumFlushEvents() < 1; i++) {
-    std::this_thread::sleep_for(std::chrono::milliseconds(50));
-    if (i > 200) {
-      FAIL() << "Timeout waiting for Flush Event Listener Results";
-    }
-  }
-
-  EXPECT_EQ(1, default_listener->GetNumFlushEvents());
-  EXPECT_EQ(1, default_listener->GetNumFlushEventsAttempted());
-  EXPECT_EQ(0, default_listener->GetNumFlushEventsFailed());
-  EXPECT_EQ(3, default_listener->GetNumFlushedRequests());
-  EXPECT_EQ(0, default_listener->GetNumFlushedRequestsFailed());
-}
-
-TEST_F(StreamLayerClientCacheMockTest,
-       FlushListenerMetricsMultipleFlushEventsInSeries) {
-  disk_cache_->Close();
-  flush_settings_.auto_flush_num_events = 2;
-  client_ = CreateStreamLayerClient();
-
-  auto default_listener = StreamLayerClient::DefaultListener();
-  client_->Enable(default_listener);
-  {
-    testing::InSequence dummy;
-
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_LOOKUP_INGEST), _, _, _, _))
-        .Times(1);
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_LOOKUP_CONFIG), _, _, _, _))
-        .Times(1);
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_GET_CATALOG), _, _, _, _))
-        .Times(1);
-    EXPECT_CALL(*network_, Send(IsPostRequest(URL_INGEST_DATA), _, _, _, _))
-        .Times(6);
-  }
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(2));
-
-  for (int i = 0, j = 1;; i++) {
-    std::this_thread::sleep_for(std::chrono::milliseconds(50));
-    if (default_listener->GetNumFlushEvents() == j) {
-      if (j == 3) {
-        break;
-      }
-      ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(2));
-      j++;
-    }
-    if (i > 400) {
-      FAIL() << "Timeout waiting for Flush Event Listener Results";
-    }
-  }
-
-  EXPECT_EQ(3, default_listener->GetNumFlushEvents());
-  EXPECT_EQ(3, default_listener->GetNumFlushEventsAttempted());
-  EXPECT_EQ(0, default_listener->GetNumFlushEventsFailed());
-  EXPECT_EQ(6, default_listener->GetNumFlushedRequests());
-  EXPECT_EQ(0, default_listener->GetNumFlushedRequestsFailed());
-}
-
-TEST_F(StreamLayerClientCacheMockTest,
-       FlushListenerMetricsMultipleFlushEventsInParallel) {
-  disk_cache_->Close();
-  flush_settings_.auto_flush_num_events = 2;
-  client_ = CreateStreamLayerClient();
-
-  auto default_listener = StreamLayerClient::DefaultListener();
-  client_->Enable(default_listener);
-  {
-    testing::InSequence dummy;
-
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_LOOKUP_INGEST), _, _, _, _))
-        .Times(1);
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_LOOKUP_CONFIG), _, _, _, _))
-        .Times(1);
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_GET_CATALOG), _, _, _, _))
-        .Times(1);
-    EXPECT_CALL(*network_, Send(IsPostRequest(URL_INGEST_DATA), _, _, _, _))
-        .Times(6);
-  }
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(6));
-
-  for (int i = 0; default_listener->GetNumFlushedRequests() < 6; i++) {
-    std::this_thread::sleep_for(std::chrono::milliseconds(50));
-    if (i > 200) {
-      FAIL() << "Timeout waiting for Flush Event Listener Results";
-    }
-  }
-
-  EXPECT_LE(3, default_listener->GetNumFlushEvents());
-  EXPECT_LE(3, default_listener->GetNumFlushEventsAttempted());
-  EXPECT_EQ(0, default_listener->GetNumFlushEventsFailed());
-  EXPECT_EQ(6, default_listener->GetNumFlushedRequests());
-  EXPECT_EQ(0, default_listener->GetNumFlushedRequestsFailed());
-}
-
-TEST_F(StreamLayerClientCacheMockTest, FlushListenerNotifications) {
-  disk_cache_->Close();
-  flush_settings_.auto_flush_num_events = 3;
-  client_ = CreateStreamLayerClient();
-  {
-    testing::InSequence dummy;
-
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_LOOKUP_INGEST), _, _, _, _))
-        .Times(1);
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_LOOKUP_CONFIG), _, _, _, _))
-        .Times(1);
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_GET_CATALOG), _, _, _, _))
-        .Times(1);
-    EXPECT_CALL(*network_, Send(IsPostRequest(URL_INGEST_DATA), _, _, _, _))
-        .Times(3);
-  }
-
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(3));
-
-  class NotificationListener : public DefaultFlushEventListener<
-                                   const StreamLayerClient::FlushResponse&> {
-   public:
-    void NotifyFlushEventStarted() override { events_started_++; }
-
-    void NotifyFlushEventResults(
-        const StreamLayerClient::FlushResponse& results) override {
-      std::lock_guard<std::mutex> lock(results_mutex_);
-      results_ = std::move(results);
-    }
-
-    const StreamLayerClient::FlushResponse& GetResults() {
-      std::lock_guard<std::mutex> lock(results_mutex_);
-      return results_;
-    }
-
-    int events_started_ = 0;
-
-   private:
-    StreamLayerClient::FlushResponse results_;
-    std::mutex results_mutex_;
-  };
-
-  auto notification_listener = std::make_shared<NotificationListener>();
-  client_->Enable(notification_listener);
-
-  for (int i = 0; notification_listener->GetResults().size() < 3; i++) {
-    std::this_thread::sleep_for(std::chrono::milliseconds(50));
-    if (i > 200) {
-      FAIL() << "Timeout waiting for Flush Event Listener Results";
-    }
-  }
-
-  EXPECT_EQ(1, notification_listener->events_started_);
-  for (auto result : notification_listener->GetResults()) {
-    ASSERT_NO_FATAL_FAILURE(PublishDataSuccessAssertions(result));
-  }
-}
-
-TEST_F(StreamLayerClientCacheMockTest, FlushDataMaxEventsDefaultSetting) {
-  {
-    testing::InSequence dummy;
-
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_LOOKUP_INGEST), _, _, _, _))
-        .Times(1);
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_LOOKUP_CONFIG), _, _, _, _))
-        .Times(1);
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_GET_CATALOG), _, _, _, _))
-        .Times(1);
-    EXPECT_CALL(*network_, Send(IsPostRequest(URL_INGEST_DATA), _, _, _, _))
-        .Times(5);
-  }
-  ASSERT_NO_FATAL_FAILURE(FlushDataOnSettingSuccessAssertions());
-}
-
-TEST_F(StreamLayerClientCacheMockTest, FlushDataMaxEventsValidCustomSetting) {
-  const int max_events_per_flush = 3;
-  disk_cache_->Close();
-  flush_settings_.events_per_single_flush = max_events_per_flush;
-  client_ = CreateStreamLayerClient();
-  {
-    testing::InSequence dummy;
-
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_LOOKUP_INGEST), _, _, _, _))
-        .Times(1);
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_LOOKUP_CONFIG), _, _, _, _))
-        .Times(1);
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_GET_CATALOG), _, _, _, _))
-        .Times(1);
-    EXPECT_CALL(*network_, Send(IsPostRequest(URL_INGEST_DATA), _, _, _, _))
-        .Times(3);
-  }
-
-  ASSERT_NO_FATAL_FAILURE(
-      FlushDataOnSettingSuccessAssertions(max_events_per_flush));
-}
-
-TEST_F(StreamLayerClientCacheMockTest, FlushDataMaxEventsInvalidCustomSetting) {
-  const int max_events_per_flush = -3;
-  disk_cache_->Close();
-  flush_settings_.events_per_single_flush = max_events_per_flush;
-  client_ = CreateStreamLayerClient();
-  {
-    testing::InSequence dummy;
-
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_LOOKUP_INGEST), _, _, _, _))
-        .Times(0);
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_LOOKUP_CONFIG), _, _, _, _))
-        .Times(0);
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_GET_CATALOG), _, _, _, _))
-        .Times(0);
-    EXPECT_CALL(*network_, Send(IsPostRequest(URL_INGEST_DATA), _, _, _, _))
-        .Times(0);
-  }
-
-  ASSERT_NO_FATAL_FAILURE(
-      FlushDataOnSettingSuccessAssertions(max_events_per_flush));
-}
-
-TEST_F(StreamLayerClientCacheMockTest, FlushSettingsTimeSinceOldRequest) {
-  disk_cache_->Close();
-  flush_settings_.auto_flush_old_events_force_flush_interval = 1;
-  client_ = CreateStreamLayerClient();
-  {
-    testing::InSequence dummy;
-
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_LOOKUP_INGEST), _, _, _, _))
-        .Times(1);
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_LOOKUP_CONFIG), _, _, _, _))
-        .Times(1);
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_GET_CATALOG), _, _, _, _))
-        .Times(1);
-    EXPECT_CALL(*network_, Send(IsPostRequest(URL_INGEST_DATA), _, _, _, _))
-        .Times(2);
-  }
-
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(2));
-
-  auto default_listener = StreamLayerClient::DefaultListener();
-  client_->Enable(default_listener);
-
-  for (int i = 0; default_listener->GetNumFlushEvents() < 1; i++) {
-    std::this_thread::sleep_for(std::chrono::milliseconds(50));
-    if (i > 20) {
-      FAIL() << "Timeout waiting for Flush Event Listener Results";
-    }
-  }
-
-  EXPECT_EQ(1, default_listener->GetNumFlushEvents());
-  EXPECT_EQ(1, default_listener->GetNumFlushEventsAttempted());
-  EXPECT_EQ(0, default_listener->GetNumFlushEventsFailed());
-  EXPECT_EQ(2, default_listener->GetNumFlushedRequests());
-  EXPECT_EQ(0, default_listener->GetNumFlushedRequestsFailed());
-}
-
-TEST_F(StreamLayerClientCacheMockTest, FlushSettingsAutoFlushInterval) {
-  disk_cache_->Close();
-  flush_settings_.auto_flush_interval = 1;
-  client_ = CreateStreamLayerClient();
-  {
-    testing::InSequence dummy;
-
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_LOOKUP_INGEST), _, _, _, _))
-        .Times(1);
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_LOOKUP_CONFIG), _, _, _, _))
-        .Times(1);
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_GET_CATALOG), _, _, _, _))
-        .Times(1);
-    EXPECT_CALL(*network_, Send(IsPostRequest(URL_INGEST_DATA), _, _, _, _))
-        .Times(2);
-  }
-
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(2));
-
-  auto default_listener = StreamLayerClient::DefaultListener();
-  client_->Enable(default_listener);
-
-  for (int i = 0; default_listener->GetNumFlushEvents() < 1; i++) {
-    std::this_thread::sleep_for(std::chrono::milliseconds(50));
-    if (i > 100) {
-      FAIL() << "Timeout waiting for Flush Event Listener Results";
-    }
-  }
-
-  EXPECT_EQ(1, default_listener->GetNumFlushEvents());
-  EXPECT_EQ(1, default_listener->GetNumFlushEventsAttempted());
-  EXPECT_EQ(0, default_listener->GetNumFlushEventsFailed());
-  EXPECT_EQ(2, default_listener->GetNumFlushedRequests());
-  EXPECT_EQ(0, default_listener->GetNumFlushedRequestsFailed());
-}
-
-TEST_F(StreamLayerClientCacheMockTest, FlushSettingsMaximumRequests) {
-  disk_cache_->Close();
-  ASSERT_EQ(flush_settings_.maximum_requests, boost::none);
-  client_ = CreateStreamLayerClient();
-  {
-    testing::InSequence dummy;
-
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_LOOKUP_INGEST), _, _, _, _))
-        .Times(1);
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_LOOKUP_CONFIG), _, _, _, _))
-        .Times(1);
-    EXPECT_CALL(*network_, Send(IsGetRequest(URL_GET_CATALOG), _, _, _, _))
-        .Times(1);
-    EXPECT_CALL(*network_, Send(IsPostRequest(URL_INGEST_DATA), _, _, _, _))
-        .Times(15);
-  }
-
-  QueueMultipleEvents(15);
-  auto response = client_->Flush().GetFuture().get();
-
-  ASSERT_EQ(15, response.size());
-  for (auto& single_response : response) {
-    ASSERT_NO_FATAL_FAILURE(PublishDataSuccessAssertions(single_response));
-  }
-  flush_settings_.maximum_requests = 10;
-  client_ = CreateStreamLayerClient();
-  ASSERT_NO_FATAL_FAILURE(MaximumRequestsSuccessAssertions(10));
-  client_ = CreateStreamLayerClient();
-  ASSERT_NO_FATAL_FAILURE(MaximumRequestsSuccessAssertions(10, 13));
-  client_ = CreateStreamLayerClient();
-  ASSERT_NO_FATAL_FAILURE(MaximumRequestsSuccessAssertions(10, 9));
-  flush_settings_.maximum_requests = 0;
-  client_ = CreateStreamLayerClient();
-  ASSERT_NO_FATAL_FAILURE(MaximumRequestsSuccessAssertions(0, 10));
 }
 
 }  // namespace
