+++ scripts/linux/psv/travis_test_psv.sh
@@ -17,7 +17,7 @@ $CPP_TEST_SOURCE_DARASERVICE_READ/olp-cpp-sdk-dataservice-read-tests \
 echo ">>> Dataservice write Test ... >>>"
 $CPP_TEST_SOURCE_DARASERVICE_WRITE/olp-cpp-sdk-dataservice-write-tests \
     --gtest_output="xml:olp-cpp-sdk-dataservice-write-tests-report.xml" \
-    --gtest_filter=-"*Online*":"StreamLayerClientCacheMockTest*"
+    --gtest_filter=-"*Online*":"StreamLayerClientCacheTest*"
 echo ">>> Integration Test ... >>>"
 $CPP_TEST_SOURCE_INTEGRATION/olp-cpp-sdk-integration-tests \
     --gtest_output="xml:olp-cpp-sdk-integration-tests-report.xml"
